Deduplicate kafka_topic
